+++ extensions/sql/asset/index/src/main/java/org/eclipse/dataspaceconnector/sql/asset/index/SqlAssetIndex.java
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Daimler TSS GmbH - Initial API and Implementation
+ *       Microsoft Corporation - added full QuerySpec support
  *
  */
 
@@ -22,7 +23,9 @@ import org.eclipse.dataspaceconnector.spi.asset.AssetIndex;
 import org.eclipse.dataspaceconnector.spi.asset.AssetSelectorExpression;
 import org.eclipse.dataspaceconnector.spi.asset.DataAddressResolver;
 import org.eclipse.dataspaceconnector.spi.persistence.EdcPersistenceException;
+import org.eclipse.dataspaceconnector.spi.query.Criterion;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
+import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.eclipse.dataspaceconnector.spi.transaction.TransactionContext;
 import org.eclipse.dataspaceconnector.spi.transaction.datasource.DataSourceRegistry;
 import org.eclipse.dataspaceconnector.spi.types.domain.DataAddress;
@@ -34,15 +37,17 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.AbstractMap;
 import java.util.HashMap;
-import java.util.LinkedList;
+import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.sql.DataSource;
 
+import static java.lang.String.format;
 import static org.eclipse.dataspaceconnector.sql.SqlQueryExecutor.executeQuery;
 
 public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolver {
+
 
     private final ObjectMapper objectMapper;
     private final DataSourceRegistry dataSourceRegistry;
@@ -67,7 +72,7 @@ public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolv
             transactionContext.execute(() -> {
                 try {
                     if (existsById(asset.getId(), connection)) {
-                        throw new EdcPersistenceException(String.format("Cannot persist. Asset with ID '%s' already exists.", asset.getId()));
+                        throw new EdcPersistenceException(format("Cannot persist. Asset with ID '%s' already exists.", asset.getId()));
                     }
 
                     executeQuery(connection, sqlAssetQueries.getSqlAssetInsertClause(),
@@ -80,7 +85,7 @@ public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolv
                         executeQuery(connection, sqlAssetQueries.getSqlPropertyInsertClause(),
                                 asset.getId(),
                                 property.getKey(),
-                                objectMapper.writeValueAsString(property.getValue()),
+                                toPropertyValue(property.getValue()),
                                 property.getValue().getClass().getName());
                     }
 
@@ -96,13 +101,6 @@ public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolv
             }
             throw new EdcPersistenceException(e.getMessage(), e);
         }
-    }
-
-    @Override
-    public void accept(AssetEntry item) {
-        Objects.requireNonNull(item);
-
-        this.accept(item.getAsset(), item.getDataAddress());
     }
 
     @Override
@@ -128,22 +126,52 @@ public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolv
     }
 
     @Override
+    public void accept(AssetEntry item) {
+        Objects.requireNonNull(item);
+
+        accept(item.getAsset(), item.getDataAddress());
+    }
+
+    @Override
     public Stream<Asset> queryAssets(AssetSelectorExpression expression) {
         Objects.requireNonNull(expression);
 
-        return getAssetStream(sqlAssetQueries.getSqlAssetListClause());
+        var criteria = expression.getCriteria();
+        var querySpec = QuerySpec.Builder.newInstance().filter(criteria)
+                .offset(0)
+                .limit(Integer.MAX_VALUE) // means effectively no limit
+                .build();
+        return queryAssets(querySpec);
     }
 
     @Override
     public Stream<Asset> queryAssets(QuerySpec querySpec) {
         Objects.requireNonNull(querySpec);
 
-        var query = String.format("%s LIMIT %s OFFSET %s",
-                sqlAssetQueries.getSqlAssetListClause(),
-                querySpec.getLimit(),
-                querySpec.getOffset());
 
-        return getAssetStream(query);
+        var conditions = querySpec.getFilterExpression().stream().map(SqlConditionExpression::new).collect(Collectors.toList());
+        var results = conditions.stream().map(SqlConditionExpression::isValidExpression).collect(Collectors.toList());
+
+        if (results.stream().anyMatch(Result::failed)) {
+            var message = results.stream().flatMap(r -> r.getFailureMessages().stream()).collect(Collectors.joining(", "));
+            throw new IllegalArgumentException(message);
+
+        }
+        var subSelects = conditions.stream().map(this::toSubSelect).collect(Collectors.toList());
+
+        var template = "%s %s LIMIT %s OFFSET %s";
+        var query = format(template, sqlAssetQueries.getSqlAssetListClause(), concatSubSelects(subSelects), querySpec.getLimit(), querySpec.getOffset());
+
+        return transactionContext.execute(() -> {
+            try (var connection = getConnection()) {
+                var params = conditions.stream().flatMap(SqlConditionExpression::toStatementParameter).collect(Collectors.toList());
+
+                List<String> ids = executeQuery(connection, this::mapAssetIds, query, params.toArray(Object[]::new));
+                return ids.stream().map(this::findById);
+            } catch (SQLException e) {
+                throw new EdcPersistenceException(e);
+            }
+        });
     }
 
     @Override
@@ -177,13 +205,7 @@ public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolv
 
         try (var connection = getConnection()) {
             var dataAddressList = transactionContext.execute(() -> executeQuery(connection, this::mapDataAddress, sqlAssetQueries.getSqlDataAddressFindByIdClause(), assetId));
-            if (dataAddressList.size() <= 0) {
-                return null;
-            } else if (dataAddressList.size() > 1) {
-                throw new IllegalStateException("Expected result set size of 0 or 1 but got " + dataAddressList.size());
-            } else {
-                return dataAddressList.iterator().next();
-            }
+            return single(dataAddressList);
         } catch (Exception e) {
             if (e instanceof EdcPersistenceException) {
                 throw (EdcPersistenceException) e;
@@ -191,37 +213,64 @@ public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolv
                 throw new EdcPersistenceException(e.getMessage(), e);
             }
         }
+    }
+
+    int mapRowCount(ResultSet resultSet) throws SQLException {
+        return resultSet.getInt(sqlAssetQueries.getCountVariableName());
+    }
+
+    AbstractMap.SimpleImmutableEntry<String, Object> mapPropertyResultSet(ResultSet resultSet) throws SQLException, ClassNotFoundException, JsonProcessingException {
+        var name = resultSet.getString(sqlAssetQueries.getAssetPropertyColumnName());
+        var value = resultSet.getString(sqlAssetQueries.getAssetPropertyColumnValue());
+        var type = resultSet.getString(sqlAssetQueries.getAssetPropertyColumnType());
+
+
+        return new AbstractMap.SimpleImmutableEntry<>(name, fromPropertyValue(value, type));
     }
 
 
     @Nullable
-    private Stream<Asset> getAssetStream(String query) {
-        try (var connection = getConnection()) {
-            return transactionContext.execute(() -> {
-                var assetIds = executeQuery(connection, this::mapAssetIds, query);
-                if (assetIds.size() <= 0) {
-                    return null;
-                }
-
-                var assets = new LinkedList<Asset>();
-                for (var assetId : assetIds) {
-                    var assetProperties = executeQuery(connection, this::mapPropertyResultSet, sqlAssetQueries.getSqlPropertyFindByIdClause(), assetId).stream().collect(Collectors.toMap(
-                            AbstractMap.SimpleImmutableEntry::getKey,
-                            AbstractMap.SimpleImmutableEntry::getValue));
-
-                    assets.add(Asset.Builder.newInstance().id(assetId).properties(assetProperties).build());
-                }
-                return assets.stream();
-            });
-
-        } catch (Exception e) {
-            if (e instanceof EdcPersistenceException) {
-                throw (EdcPersistenceException) e;
-            } else {
-                throw new EdcPersistenceException(e.getMessage(), e);
-            }
+    private DataAddress single(List<DataAddress> dataAddressList) {
+        if (dataAddressList.size() <= 0) {
+            return null;
+        } else if (dataAddressList.size() > 1) {
+            throw new IllegalStateException("Expected result set size of 0 or 1 but got " + dataAddressList.size());
+        } else {
+            return dataAddressList.iterator().next();
         }
     }
+
+    /**
+     * Deserializes a value into an object using the object mapper.
+     * Note: if type is {@code java.lang.String} simply {@code value.toString()} is returned.
+     */
+    private Object fromPropertyValue(String value, String type) throws ClassNotFoundException, JsonProcessingException {
+        var clazz = Class.forName(type);
+        if (clazz == String.class) {
+            return value;
+        }
+        return objectMapper.readValue(value, clazz);
+    }
+
+    /**
+     * Concatenates all SELECT statements on all properties into one big statement, or returns "" if list is empty.
+     */
+    private String concatSubSelects(List<String> subSelects) {
+        if (subSelects.isEmpty()) {
+            return "";
+        }
+        return format(" WHERE %s", String.join(" AND ", subSelects));
+    }
+
+    /**
+     * Converts a {@linkplain Criterion} into a dynamically assembled SELECT statement.
+     */
+    private String toSubSelect(SqlConditionExpression c) {
+        return format("%s %s %s)", sqlAssetQueries.getQuerySubSelectClause(),
+                c.getCriterion().getOperator(),
+                c.toValuePlaceholder());
+    }
+
 
     private boolean existsById(String assetId, Connection connection) {
         var assetCount = transactionContext.execute(() -> executeQuery(connection, this::mapRowCount, sqlAssetQueries.getSqlAssetCountByIdClause(), assetId).iterator().next());
@@ -236,21 +285,11 @@ public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolv
     }
 
     private DataSource getDataSource() {
-        return Objects.requireNonNull(dataSourceRegistry.resolve(dataSourceName), String.format("DataSource %s could not be resolved", dataSourceName));
+        return Objects.requireNonNull(dataSourceRegistry.resolve(dataSourceName), format("DataSource %s could not be resolved", dataSourceName));
     }
 
     private Connection getConnection() throws SQLException {
         return getDataSource().getConnection();
-    }
-
-    int mapRowCount(ResultSet resultSet) throws SQLException {
-        return resultSet.getInt(sqlAssetQueries.getCountVariableName());
-    }
-
-    AbstractMap.SimpleImmutableEntry<String, Object> mapPropertyResultSet(ResultSet resultSet) throws SQLException, ClassNotFoundException, JsonProcessingException {
-        return new AbstractMap.SimpleImmutableEntry<>(resultSet.getString(sqlAssetQueries.getAssetPropertyColumnName()),
-                objectMapper.readValue(resultSet.getString(sqlAssetQueries.getAssetPropertyColumnValue()),
-                        Class.forName(resultSet.getString(sqlAssetQueries.getAssetPropertyColumnType()))));
     }
 
     private DataAddress mapDataAddress(ResultSet resultSet) throws SQLException, JsonProcessingException {
@@ -261,5 +300,9 @@ public class SqlAssetIndex implements AssetLoader, AssetIndex, DataAddressResolv
 
     private String mapAssetIds(ResultSet resultSet) throws SQLException {
         return resultSet.getString(sqlAssetQueries.getAssetColumnId());
+    }
+
+    private String toPropertyValue(Object value) throws JsonProcessingException {
+        return value instanceof String ? value.toString() : objectMapper.writeValueAsString(value);
     }
 }